Move GraphStore implementation to its own directory
