+++ arangod/Pregel/Algos/LabelPropagation/LabelPropagation.cpp
@@ -26,7 +26,8 @@
 #include "Cluster/ServerState.h"
 #include "Pregel/Aggregator.h"
 #include "Pregel/Algorithm.h"
-#include "Pregel/Worker/GraphStore.h"
+#include "Pregel/GraphStore/GraphStore.h"
+#include "Pregel/IncomingCache.h"
 #include "Pregel/MasterContext.h"
 #include "Pregel/VertexComputation.h"
 #include "Random/RandomGenerator.h"
