+++ arangod/Pregel/GraphStore/CMakeLists.txt
@@ -1,3 +1,2 @@
 target_sources(arango_pregel PRIVATE
-  WorkerConfig.cpp
-  Worker.cpp)
+  GraphStore.cpp)
