+++ arangod/Pregel/OutgoingCache.h
@@ -27,7 +27,7 @@
 #include "Cluster/ClusterInfo.h"
 #include "VocBase/voc-types.h"
 
-#include "Pregel/Worker/GraphStore.h"
+#include "Pregel/GraphStore/GraphStore.h"
 #include "Pregel/MessageCombiner.h"
 #include "Pregel/MessageFormat.h"
 #include "Pregel/Worker/WorkerConfig.h"
